+++ packages/helper/src/features/index.ts
@@ -1,2 +1,3 @@
+export { NimbusDependenciesChecker } from "./NimbusDependenciesChecker";
 export { HoverProvider } from "./HoverProvider";
 export { Subscriptions } from "./Subscriptions";