+++ packages/react/src/atomic/Box/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 A low-level utility component that accepts styled system props to enable custom theme-aware styling
 
+## 2023-08-04 `4.1.0`
+
+#### 🎉 New features
+
+- Added `fontSize` property to the Component `Box` API. ([#183](https://github.com/TiendaNube/nimbus-design-system/pull/183) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `fontWeight` property to the Component `Box` API. ([#183](https://github.com/TiendaNube/nimbus-design-system/pull/183) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `lineHeight` property to the Component `Box` API. ([#183](https://github.com/TiendaNube/nimbus-design-system/pull/183) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `textAlign` property to the Component `Box` API. ([#183](https://github.com/TiendaNube/nimbus-design-system/pull/183) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-05-18 `4.0.0`
 
 #### 🛠 Breaking changes